==> Script/Analysis_07_Correct_Encoding_Errors.R <==
# read data
FileName <- "alljournals-2021-02-05.csv"
  
alljournals <- readr::read_csv(paste0(".\\Output\\", FileName))

# check wrongful journal names with:
# WrongOnes <- alljournals[grepl("<.*>", alljournals$journal), ]
#

# fix wrongful hex-codes
ascii <- structure(list(Hex = c("<a0>", "<a1>", "<a2>", "<a3>", "<a4>", 
                                "<a5>", "<a6>", "<a7>", "<a8>", "<a9>", "<aa>", "<ab>", "<ac>", 
                                "<ad>", "<ae>", "<af>", "<b0>", "<b1>", "<b2>", "<b3>", "<b4>", 
                                "<b5>", "<b6>", "<b7>", "<b8>", "<b9>", "<ba>", "<bb>", "<bc>", 
                                "<bd>", "<be>", "<bf>", "<c0>", "<c1>", "<c2>", "<c3>", "<c4>", 
                                "<c5>", "<c6>", "<c7>", "<c8>", "<c9>", "<ca>", "<cb>", "<cc>", 
                                "<cd>", "<ce>", "<cf>", "<d0>", "<d1>", "<d2>", "<d3>", "<d4>", 
                                "<d5>", "<d6>", "<d7>", "<d8>", "<d9>", "<da>", "<db>", "<dc>", 
                                "<dd>", "<de>", "<df>", "<e0>", "<e1>", "<e2>", "<e3>", "<e4>", 
                                "<e5>", "<e6>", "<e7>", "<e8>", "<e9>", "<ea>", "<eb>", "<ec>", 
                                "<ed>", "<ee>", "<ef>", "<f0>", "<f1>", "<f2>", "<f3>", "<f4>", 
                                "<f5>", "<f6>", "<f7>", "<f8>", "<f9>", "<fa>", "<fb>", "<fc>", 
                                "<fd>", "<fe>", "<ff>"), Actual = c(" ", "¡", "¢", "£", "¤", 
                                                                    "¥", "¦", "§", "¨", "©", "ª", "«", "¬", "SHY", "®", "¯", "°", 
                                                                    "±", "²", "³", "´", "µ", "¶", "·", "¸", "¹", "º", "»", "¼", "½", 
                                                                    "¾", "¿", "À", "Á", "Â", "Ã", "Ä", "Å", "Æ", "Ç", "È", "É", "Ê", 
                                                                    "Ë", "Ì", "Í", "Î", "Ï", "Ð", "Ñ", "Ò", "Ó", "Ô", "Õ", "Ö", "×", 
                                                                    "Ø", "Ù", "Ú", "Û", "Ü", "Ý", "Þ", "ß", "à", "á", "â", "ã", "ä", 
                                                                    "å", "æ", "ç", "è", "é", "ê", "ë", "ì", "í", "î", "ï", "ð", "ñ", 
                                                                    "ò", "ó", "ô", "õ", "ö", "÷", "ø", "ù", "ú", "û", "ü", "ý", "þ", 
                                                                    "ÿ")), row.names = c(NA, -96L), class = "data.frame")

alljournals$journal <- stringi::stri_replace_all_fixed(
  alljournals$journal, 
  ascii$Hex, 
  ascii$Actual, 
  vectorize_all = FALSE
)

# fix some of the wrongful unicodes
WRONG <- c("<U+0096>", "<U+0092>", "<U+0097>", "<U+00A0>", "<U+009A>",
           "<U+02BB>", "<U+009E>", "<U+0086>", "<U+0093>", "<U+0094>",
           "<U+1EC7>", "<U+00A8>", "<U+0450>", "<U+008A>", "<U+00AA>",
           "<U+0218>", "<U+0091>", "<U+00B4>", "<U+200B>", "<U+206F>",
           "<U+0219>", "<U+0091>", "<U+01FA>", "<U+039C>", "<U+00A1>",
           "<U+00B7>", "<U+021B>", "<U+FFFD>",  "<U+FB01>","<U+008E>",
           "<U+00AD>", "<U+00AE>", "<U+00AB>", "<U+00BB>", "<U+0099>",
           "<U+0203>", "<U+0392>", "<U+0421>", "<U+041C>", "<U+202A>",
           "<U+039A>", "<U+0441>", "<U+0410>", "<U+00A9>", "<U+200E>",
           "<U+00B3>", "<U+02BD>", "<U+1EA1>", "<U+1ECD>", "<U+1ED9>",
           "<U+04E7>", "<U+1ECC>", "<U+00A4>", "<U+0087>", "<U+E524>",
           "<U+00A2>", "<U+1EE7>", "<U+1EE9>", "<U+009C>", "<U+00BA>",
           "<U+0084>", "<U+00B2>", "<U+00B0>", "<U+021A>", "<U+03A4>",
           "<U+0391>", "<U+041E>", "<U+041F>", "<U+0413>", "<U+0422>",
           "<U+0423>", "<U+03C1>", "<U+039F>", "<U+0095>", "<U+03B3>",
           "<U+0384>", "<U+03BF>", "<U+03BD>", "<U+03C5>", "<U+2009>",
           "<U+00AC>", "<U+03CE>", "<U+03BA>", "<U+03B9>", "<U+202F>",
           "<U+00B8>", "<U+0412>", "<U+0430>", "<U+2194>")

RIGHT <- c("–", "'", "—", " ", "š",
           "ʻ", "ž", "†", "“", "”",
           "ệ", "ü", "è", "Š", "ª",
           "Ș", "'", "´", "", "",
           "ș", "'", "Ǻ", "M", "¡",
           "·", "ț", "", "fi", "Ž",
           "-", "®", "«", "»", "™",
           "ȃ", "B", "C", "M", "",
           "K", "c", "A", "©", "",
           "³", "ʽ", "ạ", "ọ", "ộ",
           "ö", "Ọ", "¤", "‡", " ",
           "¢", "ủ", "ứ", "œ", "°",
           " ", "²", "°", "Ț", "T",
           "A", "О", "П", "Г", "Т",
           "У", "ρ", "Ο", "•", "γ",
           "´", "ο", "ν", "υ", " ",
           "¬", "ώ", "κ", "ι", " ",
           "¸", "B", "a", "↔")

alljournals$journal <- stringi::stri_replace_all_fixed(
  alljournals$journal, 
  WRONG, 
  RIGHT, 
  vectorize_all = FALSE
)

# remove superfluous "</a>"-tags
alljournals$journal <- stringr::str_remove_all(alljournals$journal, "</a>")

# save corrected file
write.csv(alljournals, paste0(".\\Output\\", FileName), row.names = FALSE)
